Add typescript
